[WIP] rewrite model load
