+++ resources/shaders/demomodel.fs
@@ -1,0 +1,8 @@
+#version 410 core
+
+out vec4 fragColor;
+
+void main()
+{
+    fragColor = vec4(1.0, 0.0, 0.0, 1.0);
+}
