+++ resources/shaders/demomodel.vs
@@ -1,0 +1,16 @@
+#version 410 core
+
+layout(location = 0) in vec3 position;
+layout(location = 1) in vec2 texCoord;
+
+//out vec2 vTexCoord;
+
+uniform mat4 projection;
+uniform mat4 model;
+uniform mat4 view;
+
+void main()
+{
+    gl_Position = projection * view * model * vec4(position, 1.0);
+    //vTexCoord = texCoord;
+}
